F_temp is wrong in Flex_MB_BCS because dyn->f is wrong from flex_dyn, added derivative solvers, probably dont work

+++ FDM/FDM.c
@@ -5,7 +5,7 @@
 
 
 //Time Steppper for the Runge Kutta Method using an Explicit Integration Scheme
-matrix *step_RK_E_h(matrix *y0, matrix **Y_h, float t0, float h, Interp_function Intrpl, ODE_function odefcn_h, matrix *a, matrix *b, matrix *c, flexBody *body){
+matrix *step_RK_E_h(matrix *y0, matrix **Y_h, double t0, double h, Interp_function Intrpl, ODE_function odefcn_h, matrix *a, matrix *b, matrix *c, flexBody *body){
 
     matrix **k = (matrix **)malloc(sizeof(matrix *) * b->numCols);
